Move BoostableQuery to a separate subnamespace
