+++ src/Search/Query/TermLevel/RangeQuery.php
@@ -1,5 +1,7 @@
 <?php namespace Nord\Lumen\Elasticsearch\Search\Query\TermLevel;
 
+use Nord\Lumen\Elasticsearch\Search\Query\TermLevel\Traits\BoostableQuery;
+
 /**
  * Matches documents with fields that have terms within a certain range. The type of the Lucene query depends on the
  * field type, for string fields, the TermRangeQuery, while for number/date fields, the query is a NumericRangeQuery.
