+++ src/Search/Query/TermLevel/RegexpQuery.php
@@ -3,6 +3,7 @@
 namespace Nord\Lumen\Elasticsearch\Search\Query\TermLevel;
 
 use Nord\Lumen\Elasticsearch\Exceptions\InvalidArgument;
+use Nord\Lumen\Elasticsearch\Search\Query\TermLevel\Traits\BoostableQuery;
 
 /**
  * Class RegexpQuery
